HttpClient関連のスタブをFixture配下に移動
