+++ src/tests/Fixture/Stub/Event/AtndJsonHttpClient.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Tests\Fixture\Stub\Event;
+
+use App\Library\Http\JsonHttpClient;
+
+class AtndJsonHttpClient implements JsonHttpClient
+{
+
+    const FIXTURE_PATH = '/tests/Fixture/Response/Event/Atnd/test.json';
+
+    public function request($url)
+    {
+        return json_decode(file_get_contents(base_path() . self::FIXTURE_PATH), true);
+    }
+
+}
