+++ src/tests/Infrastructure/SocialCrawler/FacebookCountCrawlerTest.php
@@ -6,8 +6,8 @@ use App\Domain\Event\Core\EventUrl;
 use App\Domain\Rating\RatingCount\FacebookCount;
 use App\Infrastructure\SocialCrawler\RestFacebookCountCrawler;
 use PHPUnit_Framework_TestCase;
-use Tests\Infrastructure\SocialCrawler\Stub\Facebook\EmptyFacebookJsonHttpClient;
-use Tests\Infrastructure\SocialCrawler\Stub\Facebook\FacebookJsonHttpClient;
+use Tests\Fixture\Stub\Social\Facebook\EmptyFacebookJsonHttpClient;
+use Tests\Fixture\Stub\Social\Facebook\FacebookJsonHttpClient;
 
 class FacebookCountCrawlerTest extends PHPUnit_Framework_TestCase
 {
