--- src/tests/Infrastructure/SocialCrawler/Stub/Facebook/EmptyFacebookJsonHttpClient.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace Tests\Infrastructure\SocialCrawler\Stub\Facebook;
-
-use App\Library\Http\JsonHttpClient;
-
-class EmptyFacebookJsonHttpClient implements JsonHttpClient
-{
-
-    const FIXTURE_PATH = '/tests/Fixture/Response/Social/Facebook/empty.json';
-
-    public function request($url)
-    {
-        return json_decode(file_get_contents(base_path() . self::FIXTURE_PATH), true);
-    }
-
-}
